Bound pum width & height, use float for grid pos
Also add functions to get reference to types in Object (will throw)

+++ src/cursor.hpp
@@ -30,8 +30,8 @@ enum class CursorStatus : std::uint8_t
 struct CursorPos
 {
   std::uint16_t grid_num;
-  int grid_x;
-  int grid_y;
+  double grid_x;
+  double grid_y;
   int row;
   int col;
 };
