+++ src/qpaintgrid.hpp
@@ -42,7 +42,7 @@ public:
   }
   ~QPaintGrid() override = default;
   void set_size(u16 w, u16 h) override;
-  void set_pos(u16 new_x, u16 new_y) override;
+  void set_pos(double new_x, double new_y) override;
   void viewport_changed(Viewport vp) override;
   /// Process the draw commands in the event queue
   void process_events();
